chore: Refactor, move API related files to package `api`.
