+++ api/agent_models.go
@@ -1,4 +1,4 @@
-package keyfactor
+package api
 
 import "time"
