+++ api/client.go
@@ -1,4 +1,4 @@
-package keyfactor
+package api
 
 import (
 	"bytes"
