+++ examples/createStore/main.go
@@ -2,19 +2,19 @@ package main
 
 import (
 	"fmt"
-	"github.com/Keyfactor/keyfactor-go-client"
+	"github.com/Keyfactor/keyfactor-go-client/api"
 	"log"
 	"os"
 )
 
 func main() {
 	// Create a new Keyfactor client
-	clientConfig := &keyfactor.AuthConfig{
+	clientConfig := &api.AuthConfig{
 		Hostname: os.Getenv("KEYFACTOR_HOSTNAME"),
 		Username: os.Getenv("KEYFACTOR_USERNAME"),
 		Password: os.Getenv("KEYFACTOR_PASSWORD"),
 	}
-	client, err := keyfactor.NewKeyfactorClient(clientConfig)
+	client, err := api.NewKeyfactorClient(clientConfig)
 	if err != nil {
 		log.Fatal(err)
 	}
@@ -46,7 +46,7 @@ func main() {
 	properties["APIObjectId"] = "object ID"
 	properties["VaultName"] = "coolvault"
 
-	createStore := &keyfactor.CreateStoreFctArgs{
+	createStore := &api.CreateStoreFctArgs{
 		AgentId:       "keyfactor orchestrator agent ID",
 		ClientMachine: "aks_demo",
 		StorePath:     "https://coolvault.vault.azure.net/",
